Mailform: Use own ajax handler
Extbase/Fluid Ajax Widgets have bugs with caching
since they store widget configuration in session data.
Which will not be stored when the widget is cached.

Therefore we used to modified typo3 source with an unofficial patch
that put the data to the url.

This commit replaces the usage of that custom approach to be able to use
non modified typo3 source.